Begin integrating attributes class (#27)

+++ Assets/_Project/Scripts/Data/Card Data/CardEffect.cs
@@ -15,7 +15,7 @@ namespace ProjetoGatoPreto
         public EffectOperation operation = EffectOperation.ADD;
 
         [Tooltip("Value to apply. You can use negative values to subtract if this is an ADD effect.")]
-        public double value = 0.0;
+        public float value = 0.0f;
 
         [Tooltip("Event triggered when the effect is applied.")]
         public UnityEvent callEvent;
